Remove legacy endpoints and associated code
After the big release in december 2018 old endpoints and associated code
were left in place, to assure temporary backward compatibility, this
commit removes those endpoints and associated code.
