+++ web/panorama/datasets/panoramas/serialize/hal_serializer.py
