+++ web/panorama/panorama/tests/test_api_base.py
@@ -7,9 +7,8 @@ import factory
 import factory.fuzzy
 from rest_framework.test import APITestCase
 # Project
-from datasets.panoramas.v1.models import Panorama
+from datasets.panoramas.models import Panoramas
 from datasets.panoramas.tests import factories
-from panorama.management.commands.refresh_views import Command as RefreshViews
 
 
 class PanoramaApiTest(APITestCase):
@@ -20,7 +19,7 @@ class PanoramaApiTest(APITestCase):
         # Adding locations
         factories.PanoramaFactory.create(
             pano_id='PANO_1_2014',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=factory.fuzzy.FuzzyDateTime(
                 datetime.datetime(2014, 1, 1, tzinfo=UTC_TZ), force_year=2014),
             filename=factory.fuzzy.FuzzyText(length=30),
@@ -36,7 +35,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_2_2014_CLOSE',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=factory.fuzzy.FuzzyDateTime(
                 datetime.datetime(2014, 1, 1, tzinfo=UTC_TZ), force_year=2014),
             filename=factory.fuzzy.FuzzyText(length=30),
@@ -52,7 +51,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_3_2015_CLOSE',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=factory.fuzzy.FuzzyDateTime(
                 datetime.datetime(2015, 1, 1, tzinfo=UTC_TZ), force_year=2015),
             filename=factory.fuzzy.FuzzyText(length=30),
@@ -68,7 +67,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_5_2014_CLOSE_BUT_NO_CIGAR',
-            status=Panorama.STATUS.detected,
+            status=Panoramas.STATUS.detected,
             timestamp=factory.fuzzy.FuzzyDateTime(
                 datetime.datetime(2014, 1, 1, tzinfo=UTC_TZ), force_year=2014),
             filename=factory.fuzzy.FuzzyText(length=30),
@@ -84,7 +83,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_5_2015_CLOSE_BUT_NO_CIGAR',
-            status=Panorama.STATUS.to_be_rendered,
+            status=Panoramas.STATUS.to_be_rendered,
             timestamp=factory.fuzzy.FuzzyDateTime(
                 datetime.datetime(2015, 1, 1, tzinfo=UTC_TZ), force_year=2015),
             filename=factory.fuzzy.FuzzyText(length=30),
@@ -100,7 +99,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_6_2016_CLOSE',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=datetime.datetime(2016, 5, 5, tzinfo=UTC_TZ),
             filename=factory.fuzzy.FuzzyText(length=30),
             path=factory.fuzzy.FuzzyText(length=30),
@@ -115,7 +114,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_7_2016_CLOSE_BUT_NO_CIGAR',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=datetime.datetime(2016, 5, 5, tzinfo=UTC_TZ),
             filename=factory.fuzzy.FuzzyText(length=30),
             path=factory.fuzzy.FuzzyText(length=30),
@@ -130,7 +129,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_8_2017_CLOSE',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=datetime.datetime(2017, 5, 5, tzinfo=UTC_TZ),
             filename=factory.fuzzy.FuzzyText(length=30),
             path=factory.fuzzy.FuzzyText(length=30),
@@ -145,7 +144,7 @@ class PanoramaApiTest(APITestCase):
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_9_2017_CLOSE_BUT_NO_CIGAR',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=datetime.datetime(2017, 5, 5, tzinfo=UTC_TZ),
             filename=factory.fuzzy.FuzzyText(length=30),
             path=factory.fuzzy.FuzzyText(length=30),
@@ -156,11 +155,11 @@ class PanoramaApiTest(APITestCase):
             mission_distance=5.0,
             mission_year=2017,
             mission_type='bi',
-            tags=[]
+            tags=['mission-bi']
         )
         factories.PanoramaFactory.create(
             pano_id='PANO_WOZ_2018',
-            status=Panorama.STATUS.done,
+            status=Panoramas.STATUS.done,
             timestamp=datetime.datetime(2018, 5, 5, tzinfo=UTC_TZ),
             filename=factory.fuzzy.FuzzyText(length=30),
             path=factory.fuzzy.FuzzyText(length=30),
@@ -174,5 +173,4 @@ class PanoramaApiTest(APITestCase):
             tags=[]
         )
 
-        RefreshViews().refresh_views()
         time.sleep(2)
