+++ web/panorama/panorama/tests/test_api_thumbnail.py
@@ -7,8 +7,9 @@ import factory
 import factory.fuzzy
 # Project
 from . test_api_base import PanoramaApiTest
-from datasets.panoramas.v1.models import Panorama
+from datasets.panoramas.models import Panoramas
 from datasets.panoramas.tests import factories
+
 
 class ThumbnailApiTest(PanoramaApiTest):
 
@@ -24,14 +25,14 @@ class ThumbnailApiTest(PanoramaApiTest):
         self.assertEqual(response.data['heading'], 270)
 
     def test_get_thumbnail_returns_jpg(self):
-        response = self.client.get('/panorama/thumbnail/?lat=52.377956&lon=4.897070&radius=1000',
+        response = self.client.get('/panorama/thumbnail/?lat=52.377956&lon=4.897070&radius=200',
                                    follow=False,
                                    HTTP_ACCEPT='image/jpeg')
         self.assertEqual(response.status_code, 302)
         self.assertEqual(response['Location'], 'http://testserver/panorama/thumbnail/PANO_9_2017_CLOSE_BUT_NO_CIGAR/?heading=270')
 
     def test_get_thumbnail_returns_json_with_qparams(self):
-        response = self.client.get('/panorama/thumbnail/?lat=52.377956&lon=4.897070&radius=1000&width=600',
+        response = self.client.get('/panorama/thumbnail/?lat=52.377956&lon=4.897070&radius=200&width=600',
                                    follow=False,
                                    HTTP_ACCEPT='image/jpeg')
         self.assertEqual(response.status_code, 302)
